[REFACTOR] file structure
