+++ app/src/main/java/com/example/project_flow_android/di/module/ChatViewModelModule.kt
@@ -1,4 +1,4 @@
-package com.example.project_flow_android.di
+package com.example.project_flow_android.di.module
 
 import com.example.project_flow_android.viewmodel.ChatViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
